Reorganizes the packages so that the common classes are all in the common package.
