+++ ElevatorSim/src/elevatorsim/tests/FileParserTest.java
@@ -8,9 +8,9 @@ import java.util.HashMap;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
+import elevatorsim.common.constants.Direction;
 import elevatorsim.common.requests.ElevatorRequest;
-import elevatorsim.constants.Direction;
-import elevatorsim.util.FileParser;
+import elevatorsim.common.util.FileParser;
 /**
  * Test cases for the file parser
  * @author Rahul Anilkumar
